requirements.txt
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-telegram-bot==10.0.2
requests==2.18.4
urllib3==1.22
